Major changes to getting probability
Added new class CoefficentCalculator, this defines an object which
stores the award title, the name of the organization, and the
coefficient for that organization and award.

The OscarGenieDriver now has a new method readCoefficents which scans a
single CSV file at the beginning of launching the program, this reads
the coefficient for every organization and every award and adds them to
a CoefficentCalculator object, then the objects are stored in a hashmap.

The get probability method now doesn’t read the coefficient from the
file, it now reads them from the CoeffMap defined in the readCoefficent
method,

Fixed spelling errors in CSV files, removed coefficients from
calculations files and changed Golden Globe awards to define if the
award was for drama or musical/comedy.